test: share the ToolContext builders and the public-surface host store
makeCtx had 22 definitions across the agent SDK's tests in 12 shapes. Two
conventions accounted for 13 of them and differ only in call style, so
tests/helpers/tool-context.ts provides both — toolContext({...}) rooted at
tmpdir and toolContextIn(cwd, {...}) — plus gatedToolContext for the two fs
suites that pre-arm the read-before-write gate. The two planmode/worktree
suites that need a real process env now wrap the shared builder in three lines
instead of carrying a ninth near-copy. The nine genuine one-offs are left as
they are.

On the maestro side, hostStore had four copies (three identical, one adding
the optional deleteSession seam). tests/helpers/host-store.ts holds it, and
imports from the bare 'silver-core-maestro-sdk' specifier rather than
../../src — those suites exist to prove a host can implement the store seam
from the published surface alone, and pointing them at source types would
quietly cost them that.

3237 agent SDK tests green, 421 maestro green.
