+++ projects/silver-core-maestro-sdk/tests/assembly.test.ts
@@ -12,41 +12,9 @@ import {
   LedgerDriver,
   SESSION_STATES,
   type DriverEvent,
-  type LedgerStore,
-  type QueryRecord,
-  type SessionFilter,
-  type SessionRecord,
   type SessionState,
 } from 'silver-core-maestro-sdk';
-
-/** The host's own storage battery, written against the public seam only. */
-function hostStore(): LedgerStore {
-  const sessions = new Map<string, SessionRecord>();
-  const queries: QueryRecord[] = [];
-  return {
-    async putSession(r) {
-      sessions.set(r.id, { ...r });
-    },
-    async getSession(id) {
-      const r = sessions.get(id);
-      return r === undefined ? null : { ...r };
-    },
-    async listSessions(filter?: SessionFilter) {
-      let all = [...sessions.values()];
-      if (filter?.states !== undefined) all = all.filter((s) => filter.states!.includes(s.state));
-      if (filter?.dueBefore !== undefined) {
-        all = all.filter((s) => s.nextRunAt !== null && s.nextRunAt <= filter.dueBefore!);
-      }
-      return all.map((s) => ({ ...s }));
-    },
-    async appendQuery(r) {
-      queries.push({ ...r });
-    },
-    async listQueries(sessionId) {
-      return queries.filter((q) => q.sessionId === sessionId).map((q) => ({ ...q }));
-    },
-  };
-}
+import { hostStore } from './helpers/host-store.js';
 
 beforeEach(() => {
   vi.useFakeTimers();
