+++ projects/silver-core-sdk/tests/tools-task.test.ts
@@ -6,7 +6,6 @@
  */
 
 import { describe, expect, it } from 'vitest';
-import { tmpdir } from 'node:os';
 
 import type { ToolContext, ToolResultPayload } from '../src/internal/contracts.js';
 import { AbortError } from '../src/errors.js';
@@ -19,17 +18,7 @@ import {
 } from '../src/tools/task.js';
 import { createBuiltinTools } from '../src/tools/index.js';
 import * as D from '../src/tools/descriptions.js';
-
-function makeCtx(overrides: Partial<ToolContext> = {}): ToolContext {
-  return {
-    cwd: tmpdir(),
-    additionalDirectories: [],
-    env: {},
-    signal: new AbortController().signal,
-    debug: () => {},
-    ...overrides,
-  };
-}
+import { toolContext as makeCtx } from './helpers/tool-context.js';
 
 function text(r: ToolResultPayload): string {
   return String(r.content);
